haku/console: trim shipped state from PLAN/TODO, dedup test app-construction
Knowledge-hygiene + refactor cleanup after the launch-routine work shipped.

Docs (forward-looking only / SSOT):
- PLAN.md "agent-authored console": drop the now-shipped current-state prose
  (the haku-console namespace split, trace/capability tiers, launch button all
  landed); keep only what's still to build (1a richer declarative UI +
  executions panel, 1b free-form iframe + north star). Current contract ->
  console/README.md; boundary doctrine + free-form design -> the design doc.
- TODO.md Console: remove the shipped launch-button entry (with its stale
  `POST /api/launch` / mitmproxy-egress claims); fold the one-in-flight guard
  into the still-pending "recent executions" item (gated on a routine-runs
  listing API that doesn't exist -> deep-link interim; adopt anthropic SDK then).
- README.md: collapse the "## Boundary" section whose first bullet triplicated
  the Trust-boundary para + Perimeter/deploy; keep the unique content-vs-look point.

Code/tests:
- deps.py: factor the two `app.state` typed accessors (GitStateDep, SettingsDep)
  out of trace.py/capabilities.py into one shared module.
- conftest.py: add a `make_client(**settings_overrides)` factory fixture;
  refactor the `client`/`cap_client` fixtures and the unconfigured-503 test to
  use it, collapsing the triplicated create_app+TestClient+model_copy boilerplate.

bbr test //haku/console/... green; pre-commit clean.
